requirements.txt
alembic==0.9.5
click==6.7
gitdb2==2.0.2
GitPython==2.1.5
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.3.1
python-dateutil==2.6.1
python-editor==1.0.3
six==1.10.0
smmap2==2.0.3
SQLAlchemy==1.1.13
